V0.6.5 (#184)
* Masterscript: patch benchmarking

* Templates for YCSB, HammerDB and Benchbase

* Configuration: servicename_sut for fixed service

* Config: New config format and update of connection properties

* Example workload

* Dummy DBMS (PostgreSQL used as a placeholder for no DBMS has to be installed)

* Requirement: Redis client

* Dummy DBMS (PostgreSQL used as a placeholder for no DBMS has to be installed)

* Dummy DBMS busybox

* Dummy DBMS busybox sleep endlessly

* Busybox does not show host infos

* DBMSBenchmarker template new version

* Cluster: pod_log_exists()

* Cluster: pod_log_exists()

* DBMSBenchmarker template new version

* DBMSBenchmarker template new version, default is no-dev-env

* Always pull latest images of benchmarker components (dev setting)

* Images: DBMSBenchmarker, monitoring, dashboard

* Images: DBMSBenchmarker, monitoring, dashboard current versions

* k8s templates cleaned

* Images: Benchbase, HammerDB, YCSB

* Images: TPC-H

* DBMSBenchmarker: Show start also in epoch seconds

* DBMSBenchmarker: bexhoma_end_epoch_computed

* DBMSBenchmarker: Dockerfile template

* DBMSBenchmarker: bexhoma_end_epoch_computed

* DBMSBenchmarker: bexhoma_end_epoch_computed, remove " [s]" at the end

* Experiment: Add read benchmark files to connection config

* TPCTC23: Add some notes

* TPCTC23: Add some results

* YCSB: Example

* k8s templates cleaned - no privileged mode

+++ experiments/tpch/MonetDB/initschema-tpch.sql
@@ -1,15 +1,15 @@
 
 -- sccsid:     @(#)dss.ddl  2.1.8.1
-create table nation  ( n_nationkey  bigint not null,
+create table nation  ( n_nationkey  integer not null,
                             n_name       char(25) not null,
-                            n_regionkey  bigint not null,
+                            n_regionkey  integer not null,
                             n_comment    varchar(152));
 
-create table region  ( r_regionkey  bigint not null,
+create table region  ( r_regionkey  integer not null,
                             r_name       char(25) not null,
                             r_comment    varchar(152));
 
-create table part  ( p_partkey     bigint not null,
+create table part  ( p_partkey     integer not null,
                           p_name        varchar(55) not null,
                           p_mfgr        char(25) not null,
                           p_brand       char(10) not null,
@@ -19,31 +19,31 @@ create table part  ( p_partkey     bigint not null,
                           p_retailprice decimal(16,2) not null,
                           p_comment     varchar(23) not null );
 
-create table supplier ( s_suppkey     bigint not null,
+create table supplier ( s_suppkey     integer not null,
                              s_name        char(25) not null,
                              s_address     varchar(40) not null,
-                             s_nationkey   bigint not null,
+                             s_nationkey   integer not null,
                              s_phone       char(15) not null,
                              s_acctbal     decimal(16,2) not null,
                              s_comment     varchar(101) not null);
 
-create table partsupp ( ps_partkey     bigint not null,
-                             ps_suppkey     bigint not null,
+create table partsupp ( ps_partkey     integer not null,
+                             ps_suppkey     integer not null,
                              ps_availqty    integer not null,
                              ps_supplycost  decimal(16,2)  not null,
                              ps_comment     varchar(199) not null );
 
-create table customer ( c_custkey     bigint not null,
+create table customer ( c_custkey     integer not null,
                              c_name        varchar(25) not null,
                              c_address     varchar(40) not null,
-                             c_nationkey   bigint not null,
+                             c_nationkey   integer not null,
                              c_phone       char(15) not null,
                              c_acctbal     decimal(16,2)   not null,
                              c_mktsegment  char(10) not null,
                              c_comment     varchar(117) not null);
 
-create table orders  ( o_orderkey       bigint not null,
-                           o_custkey        bigint not null,
+create table orders  ( o_orderkey       integer not null,
+                           o_custkey        integer not null,
                            o_orderstatus    char(1) not null,
                            o_totalprice     decimal(16,2) not null,
                            o_orderdate      date not null,
@@ -52,9 +52,9 @@ create table orders  ( o_orderkey       bigint not null,
                            o_shippriority   integer not null,
                            o_comment        varchar(79) not null);
 
-create table lineitem ( l_orderkey    bigint not null,
-                             l_partkey     bigint not null,
-                             l_suppkey     bigint not null,
+create table lineitem ( l_orderkey    integer not null,
+                             l_partkey     integer not null,
+                             l_suppkey     integer not null,
                              l_linenumber  integer not null,
                              l_quantity    decimal(16,2) not null,
                              l_extendedprice  decimal(16,2) not null,
@@ -68,4 +68,7 @@ create table lineitem ( l_orderkey    bigint not null,
                              l_shipinstruct char(25) not null,
                              l_shipmode     char(10) not null,
                              l_comment      varchar(44) not null);
+
+
+
